requirement.txt
appdirs==1.4.3
Django==1.10.6
facebook-sdk==2.0.0
oauthlib==2.0.1
packaging==16.8
pkg-resources==0.0.0
PyJWT==1.4.2
pyparsing==2.2.0
python-openid==2.2.5
requests==2.13.0
requests-oauthlib==0.8.0
six==1.10.0
social-auth-app-django==1.1.0
social-auth-core==1.2.0
